+++ instrumentation/opentelemetry-instrumentation-pika/tests/test_utils.py
@@ -13,9 +13,15 @@
 # limitations under the License.
 from unittest import TestCase, mock
 
+from pika.channel import Channel
+from pika.spec import Basic, BasicProperties
+
 from opentelemetry.instrumentation.pika import utils
-from opentelemetry.semconv.trace import SpanAttributes
-from opentelemetry.trace import Span, Tracer
+from opentelemetry.semconv.trace import (
+    MessagingOperationValues,
+    SpanAttributes,
+)
+from opentelemetry.trace import Span, SpanKind, Tracer
 
 
 class TestUtils(TestCase):
@@ -32,12 +38,15 @@ class TestUtils(TestCase):
         channel = mock.MagicMock()
         properties = mock.MagicMock()
         task_name = "test.test"
+        span_kind = mock.MagicMock(spec=SpanKind)
         get_value.return_value = None
         ctx = mock.MagicMock()
-        _ = utils._get_span(tracer, channel, properties, task_name, ctx)
+        _ = utils._get_span(
+            tracer, channel, properties, task_name, span_kind, ctx
+        )
         generate_span_name.assert_called_once()
         tracer.start_span.assert_called_once_with(
-            context=ctx, name=generate_span_name.return_value
+            context=ctx, name=generate_span_name.return_value, kind=span_kind
         )
         enrich_span.assert_called_once()
 
@@ -54,9 +63,12 @@ class TestUtils(TestCase):
         channel = mock.MagicMock()
         properties = mock.MagicMock()
         task_name = "test.test"
+        span_kind = mock.MagicMock(spec=SpanKind)
         get_value.return_value = True
         ctx = mock.MagicMock()
-        span = utils._get_span(tracer, channel, properties, task_name, ctx)
+        span = utils._get_span(
+            tracer, channel, properties, task_name, span_kind, ctx
+        )
         self.assertEqual(span, None)
         generate_span_name.assert_not_called()
         enrich_span.assert_not_called()
@@ -158,3 +170,117 @@ class TestUtils(TestCase):
                 ),
             ],
         )
+
+    @mock.patch("opentelemetry.instrumentation.pika.utils._get_span")
+    @mock.patch("opentelemetry.propagate.extract")
+    @mock.patch("opentelemetry.trace.use_span")
+    def test_decorate_callback(
+        self,
+        use_span: mock.MagicMock,
+        extract: mock.MagicMock,
+        get_span: mock.MagicMock,
+    ) -> None:
+        callback = mock.MagicMock()
+        mock_task_name = "mock_task_name"
+        tracer = mock.MagicMock()
+        channel = mock.MagicMock(spec=Channel)
+        method = mock.MagicMock(spec=Basic.Deliver)
+        properties = mock.MagicMock()
+        mock_body = b"mock_body"
+        decorated_callback = utils._decorate_callback(
+            callback, tracer, mock_task_name
+        )
+        retval = decorated_callback(channel, method, properties, mock_body)
+        extract.assert_called_once_with(
+            properties.headers, getter=utils._pika_getter
+        )
+        get_span.assert_called_once_with(
+            tracer,
+            channel,
+            properties,
+            span_kind=SpanKind.CONSUMER,
+            task_name=mock_task_name,
+            ctx=extract.return_value,
+            operation=MessagingOperationValues.RECEIVE,
+        )
+        use_span.assert_called_once_with(
+            get_span.return_value, end_on_exit=True
+        )
+        callback.assert_called_once_with(
+            channel, method, properties, mock_body
+        )
+        self.assertEqual(retval, callback.return_value)
+
+    @mock.patch("opentelemetry.instrumentation.pika.utils._get_span")
+    @mock.patch("opentelemetry.propagate.inject")
+    @mock.patch("opentelemetry.context.get_current")
+    @mock.patch("opentelemetry.trace.use_span")
+    def test_decorate_basic_publish(
+        self,
+        use_span: mock.MagicMock,
+        get_current: mock.MagicMock,
+        inject: mock.MagicMock,
+        get_span: mock.MagicMock,
+    ) -> None:
+        callback = mock.MagicMock()
+        tracer = mock.MagicMock()
+        channel = mock.MagicMock(spec=Channel)
+        method = mock.MagicMock(spec=Basic.Deliver)
+        properties = mock.MagicMock()
+        mock_body = b"mock_body"
+        decorated_basic_publish = utils._decorate_basic_publish(
+            callback, channel, tracer
+        )
+        retval = decorated_basic_publish(
+            channel, method, mock_body, properties
+        )
+        get_current.assert_called_once()
+        get_span.assert_called_once_with(
+            tracer,
+            channel,
+            properties,
+            span_kind=SpanKind.PRODUCER,
+            task_name="(temporary)",
+            ctx=get_current.return_value,
+            operation=None,
+        )
+        use_span.assert_called_once_with(
+            get_span.return_value, end_on_exit=True
+        )
+        get_span.return_value.is_recording.assert_called_once()
+        inject.assert_called_once_with(properties.headers)
+        callback.assert_called_once_with(
+            channel, method, mock_body, properties, False
+        )
+        self.assertEqual(retval, callback.return_value)
+
+    @mock.patch("opentelemetry.instrumentation.pika.utils._get_span")
+    @mock.patch("opentelemetry.propagate.inject")
+    @mock.patch("opentelemetry.context.get_current")
+    @mock.patch("opentelemetry.trace.use_span")
+    @mock.patch("pika.spec.BasicProperties.__new__")
+    def test_decorate_basic_publish_no_properties(
+        self,
+        basic_properties: mock.MagicMock,
+        use_span: mock.MagicMock,
+        get_current: mock.MagicMock,
+        inject: mock.MagicMock,
+        get_span: mock.MagicMock,
+    ) -> None:
+        callback = mock.MagicMock()
+        tracer = mock.MagicMock()
+        channel = mock.MagicMock(spec=Channel)
+        method = mock.MagicMock(spec=Basic.Deliver)
+        mock_body = b"mock_body"
+        decorated_basic_publish = utils._decorate_basic_publish(
+            callback, channel, tracer
+        )
+        retval = decorated_basic_publish(channel, method, body=mock_body)
+        basic_properties.assert_called_once_with(BasicProperties, headers={})
+        get_current.assert_called_once()
+        use_span.assert_called_once_with(
+            get_span.return_value, end_on_exit=True
+        )
+        get_span.return_value.is_recording.assert_called_once()
+        inject.assert_called_once_with(basic_properties.return_value.headers)
+        self.assertEqual(retval, callback.return_value)