+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opentelemetry-util-http` no longer contains an instrumentation entrypoint and will not be loaded
 automatically by the auto instrumentor.
 ([#745](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/745))
+- `opentelemetry-instrumentation-pika` Bugfix use properties.headers. It will prevent the header injection from raising.
+  ([#740](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/740))
 
 ## [1.6.0-0.25b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.6.0-0.25b0) - 2021-10-13
 ### Added